backend/requirements.txt
fastapi>=0.111.0
uvicorn>=0.29.0
langchain>=0.2.0
langchain-groq>=0.1.6
langgraph>=0.1.1
yfinance>=0.2.40
httpx>=0.27.0
python-dotenv>=1.0.1
pydantic>=2.7.1
langsmith>=0.1.81
slowapi>=0.1.9
